+++ src/lib.js
@@ -1,23 +1,25 @@
-import Accordion from "./components/accordion.vue"
-import AccordionPanel from "./components/accordion-panel.vue"
-import Autocomplete from "./components/autocomplete.vue"
+import Accordion from './components/accordion.vue'
+import AccordionPanel from './components/accordion-panel.vue'
+import Autocomplete from './components/autocomplete.vue'
 import Confirm from './components/confirm.vue'
 import DateInput from './components/date-input.vue'
 import Datepicker from './components/datepicker.vue'
-import FormFileButton from "./components/form-file-button.vue"
+import FormFileButton from './components/form-file-button.vue'
 import FormSelect from './components/form-select.vue'
 import FormSwitch from './components/form-switch.vue'
 import MessageToast from './components/message-toast.vue'
-import Modal from "./components/modal.vue"
+import Modal from './components/modal.vue'
 import Pagination from './components/pagination.vue'
 import PasswordInput from './components/password-input.vue'
-import SimpleTree from "./components/simple-tree.vue"
-import Sizable from "./components/sizable.vue"
-import Slider from "./components/slider.vue"
+import SimpleTree from './components/simple-tree.vue'
+import Sizable from './components/sizable.vue'
+import Slider from './components/slider.vue'
 import Sortable from './components/sortable.vue'
-import Spinner from "./components/spinner.vue"
-import SubmitButton from "./components/submit-button.vue"
-import Tabs from "./components/tabs.vue"
+import Spinner from './components/spinner.vue'
+import SubmitButton from './components/submit-button.vue'
+import Tabs from './components/tabs.vue'
+import VFocus from './directives/focus'
+import VFloatingLabel from './directives/floatingLabel'
 
 export {
     Accordion,
@@ -39,5 +41,7 @@ export {
     Sortable,
     Spinner,
     SubmitButton,
-    Tabs
+    Tabs,
+    VFocus,
+    VFloatingLabel
 }